requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
python-dotenv==0.18.0
python-slugify==5.0.2
python-telegram-bot==13.6
pytube==10.8.5
pytz==2021.1
six==1.16.0
text-unidecode==1.3
tornado==6.1
tzlocal==2.1
